Initial set up.
Here two solutions are implemented which both have merit. The locks
one is a sort of naive solution with locks on all nodes and a
shared locked list of active nodes. It has also been optimized
some.

The second solution is based on a research paper
https://www.cs.cmu.edu/~guyb/papers/BBS15.pdf and runs the algorithm
in phases using barriers. This eliminates the need for any locking.
It has been optimized a bit and now uses transactional memory, but
that can be exchanged to atomics (original solition). It needs
the same optimizations as the locks one to have any chance.